feat/tableCalc : 계산 함수 다이얼로그 리팩토링
